Found the error for resizing the window past the minimum limit

+++ DnDAssistant/DnDAssistant.Wpf/ViewModels/WindowViewModel.cs
@@ -64,12 +64,12 @@ namespace DnDAssistant.Wpf
         /// <summary>
         /// The minimum height of the window
         /// </summary>
-        public double MinimumWindowHeight { get; set; } = 400;
+        public int MinimumWindowHeight { get; set; } = 400;
 
         /// <summary>
         /// The minimum width of the window
         /// </summary>
-        public double MinimumWindowWidth { get; set; } = 750;
+        public int MinimumWindowWidth { get; set; } = 750;
 
         /// <summary>
         /// The size of the resizeborder around the window
